Adjust text size of live meter text when large font is on.
This stops the units from being cropped in large text mode,
and fixes an accessibility problem.

Also increases the card min height to 260dp, which will help
smaller screens and large fonts to crop less when there are
too many lines of descriptive text. The previous value, 240dp,
was derived from adding heights in this spec:
https://folio.googleplex.com/scinotebook/Mock/Sensor%20card/Latest#%2FGraphic%20Card_spec.jpg
It was not itself spec'ed out. Adding another 20dp of height
allows us buffer for 3 lines of text or small screens. Tested
on a Moto X and a Nexus 6, this seems reasonable.

BUG=28726717

Change-Id: I7ebaa891498505723998747e96c1ce65d8a83926