First set of changes toward new Discoverer API
For the new manage devices screen, ExternalSensorDiscoverers can't
assume that they're talking directly to a preference screen, so
this creates a ConnectableSensor model class for a ListAdapter
to be used later.

Also, corrects a bug in ScalarInputSpec, which had the possibility
of confusing two sensors from different providers with the same
address.

Change-Id: Idcdba6418ca0d5b329b3922a4af789de6a9c8b0f