Allow sensors to display custom number of values after the decimal.
SensorApperance can take info about the number of values after the
decimal to display, and is now in charge of a sensor's value's
NumberFormat. Default is unchanged, but barometer uses the new
version.

Stats views and meter mode big value become "resizableTextViews",
and resize based on the text inside of them.

BUG=31154986,31513535

Change-Id: I913b142f41b7e4fd2daee3081dfc506206297d55