Add option to delay stopping a sensor to avoid churn.
[For review from a design and functionality point of view.]

[If we like it, should I put it behind a testing option?]

Another option for the start/stop churn for API sensors.

This option is close to Justin's suggestion: when we ask
sensors to stop, we wait for N seconds to see if they are
re-started, before finally stopping.

RecorderController is the object that will eventually move
to the sensor service, so this is the right direction to
move toward the service solution.

BUG=31670293

Change-Id: If0600c56b872d970d2ba6a1be7ad925daa726c40