Reduce work done on main thread when adding a new point to ChartView.
Removes an extra call to transformPath when adding a new data point.
transformPath is already called when the external axis updates, which
is pretty frequently, so it is not necessary to do it on every new
data point in order to have a smooth UI for the user.

This makes sensors with high rates of data draw more smoothly.

BUG=32501948

Change-Id: I3d85c0a9cf1a6f585a3950090b62b47eafd4d506