Updates to views and layouts to support large screens.
RunReview, Exp Details and Project Details now have a max width,
as does the Learn More page.

In addition: the Meter Mode stats view was changed so that it
can grow into large screens and shrink to small screens without
issue. The Project Details and Exp. Details pages had their
description / archived section updated to allow for an element
overlap per new mocks, and to center the text and archived
indicators. Because of the overlap, they can no longer be
transparent when archived, so now archived color state is
handled slightly differently.

This does not address the tutorial page.

BUG=33580330,33588458

Change-Id: Iae8836d93039b34c1d0e61194a57ee8be3126428