Fix refocus when cancling crop and zoomed to area outside of the crop.
This fixes the zoom issue but the seekbar thumb may be drawn in the
wrong place if it was way offscreen during the last crop and we are
resuming cropping. It doesn't impact where the RunReviewOverlay draws
the vertical line, the selected crop point, or the ability for the user
to touch the crop bars. Even TalkBack reads the misplaced crop seekbar
correctly, so it feels lower priority and is tricky to debug.
I'll tackle it separately.

BUG=34834694

Change-Id: I0d57703aa823e5b633b2103a4bb521f9616cb840