Move spinner management out of RecordFragment
The activity that manages RecordFragment should control how to
respond to recording events and experiment changes, rather than
allowing RecordFragment to muck about in the activity's view
hierarchy innards.

This is the first step of several; next we'll give the actionbar
similar treatment.

The end goal is making it easier to embed RecordFragment in
a different activity (for the new paned activity, for example)
without making so many assumptions about the activity's UI

BUG=33170199

Change-Id: I9bd90cc03dfdf130cef433fa9ec27e41bffaa04a