Move ApplicationLabel out of the Label class.
This prepares for a move to filemetadata.Label from metadata.Label.
ApplicationLabels are not part of the file-based structure so they
need to be separated from metadata.Label before wrapper classes
can be made for metadata.Label to filemetadata.Label.

Also deletes Run.java, which was meant to be deleted in the last CL.

BUG=36287334

Change-Id: I43f8535e374bf3527b8966a3a1925bcedabf9daa