Remove all label getters, instead using filemetadata objects for access.
This moves labels into experiments, and also removes the DataController
method to get labels for a trial, instead just getting the trial.

The next step is to have all *writes* to labels use the Experiment/Trial
objects, rather than writing individual labels. This use of a nested
structure for write will make it easier to transition to the file-based
system, where the Experiment/Trial/Labels are all saved in one file.

BUG=36287334

Change-Id: I1118e8f3529418ccd39eb706a064906aee0dbce2