Create classes for managing files of Experiments and SharedMetadata.
ExperimentCache, FileMetadataManager and SharedMetadataManager and
their tests.

FileMetadataManager is incomplete, but included in this change to
demonstrate how the other classes will be used.

Also changes updateLastUsedExperiment to setLastUsedExperiment, and
does not save the experiment when doing so. The saving was unneeded
in the current implementation and the FileMetadata implementation
is simplified if saving is removed.

BUG=36287334

Change-Id: I681cd8a46461dd116affb6b5de5f216be17a5889