Use ExperimentOverview instead of Experiment to display experiment list.
ExperimentOverview will be kept in a local part of the file-based
storage system and not backed up to Drive. Loading all Experiments is
a waste of resources just to view the list.

NOTE: this breaks the ability to show images in the ExperimentList page.
I can fix this for the database-based storage system if desired, but
would prefer to wait until the file-based system is implemented.

BUG=36287334
Change-Id: I32cd158fecfa116a46e30a6a1d1504277a6d0dc2