Move trials into Experiments.
This breaks several parts of Trials which will be fixed when
trials and experiments are accessed from files, in the next
big change:

This breaks cropping because we no longer edit ApplicationLabels.
This also breaks trial count in the Experiment List.
Finally, this breaks *saving* trials back to the database.

BUG=36287334

Change-Id: Ieb0ec3b2137e8345dadab8bd793e2399e9657cb5