Create a DataService.
Over time, the goal is to use this service for global data and services,
rather than getting those from a static AppSingleton that any activity
can access.

To begin with, activities and fragments can bind to this service, which
under the hood, serves up the same old AppSingleton.  But once everyone
is using DataService, we can eliminate the global AppSingleton in favor
of one managed directly by the service.

Otherwise, hoping that docs directly on the class answer questions.

Change-Id: I842471c44a151021b2c07541ff144946e856e8ec