Fix integration tests.  Delete MetadataController.  Fix crash.
Delete MetadataController.  We no longer have a concept of a globally "active"
experiment that can change during the lifetime of an observe fragment, so replace
the complexity of MetadataController with a simple method to load the most recent
experiment, or create one if there is none.

In the process, delete some integration tests that were failing.

And fix a crash.

BUG=62303726
BUG=62573127

Change-Id: I9241c5a3e782d07803d530c4fb1aadd805b5200f