Combine meter and graph modes.
Moves the short description into the "learn more" page.
Removes the whole meter mode and puts the icon and live data
on the graph side.

Updates the stats view to use smaller fonts and only one view
across the whole app to be more similar to the spec.

Some tweaks in color for stats overlay in graph, etc still needed.

BUG=62998939
Change-Id: I44a62dd21100579a3460421e72649ffb818e3cf3