RecorderController is a main-process shared object
The goal is that RecorderController lives in the main process, and
can be accessed by activities and services; we'll have a global state
Behavior<RecordingStatus> that will be updated by RecorderController,
and can be subscribed by others.

There should be no behavior changes in this CL, just removing an
unused layer of indirection.

BUG=62342044

Change-Id: I1476f3dd87911fabdd7c25e02e49259c7d189beb