Keep RunReview seekbar from disappearing.
This keeps the seekbar from disappearing permanantly,
although occasionally after rotate it is missing but it
can be retrieved by touching the axis, an add note dot,
zooming, or hitting the sonification button.

This gets the seekbar to a "good enough" point, although
further work could be done to keep it from disappearing
after some rotates (< 25% of rotates on my 5x).

BUG=63351103
Change-Id: I067c109fb04f0bb1fa2e488cc1e7295e32837097