Add a grabber button to the tool drawer which works in a11y modes.
The grabber is shorter than the a11y min height of 48dp, however
it sounds like that's OK by the a11y team because of how it will
be used.

To fit the grabber, the tablayout height was decreased to the min
a11y size so that the whole bar need only be increased by 2dp.

BUG=65261330

Change-Id: I3d1f3de214867fd06bfa7f7a97123058223664fb