Don't try to chop to half-size if the JPG is rotated.
Samsung Edge, and some other Samsung phones, it appears,
store the captured JPEGs rotated, with EXIF metadata defining
their orientation.  Without unrotating the image, chopping
them results in weirdness.

I've written code that first rotates them, and then
chops them, but that OOMs.  It should be possible to do
the rotation and crop as one matrix operation, but I
don't seem to have quite the cleverness I need to get that
done right now.

This solution is not-quite-right, but has the advantage
of not being very, very wrong.

BUG=67335604

Change-Id: I72912f679d41b864791f65ce470e05999325c6cd