Use different SharedPreferences files for different accounts. Copy old preferences to the current account.
Modified SimpleBackupAgent to backup the default SharedPreferences and all account specific SharePreferences.
Modified SettingsFragment to set the SharedPreferencesName on the fragment's PreferenceManager to the account specific SharedPreferences name.
Removed registerOnSharedPreferenceChangeListener in AppSingleton.getSensorRegistry.
Modified ExperimentListFragment to use the account specific SharedPreferences.
Modified AccountUtils.removeOtherAccounts to look at all account specific SharedPreferences snd remove ones that belong to an account that is no longer present.

Added code to MainActivity to register account-based preferences that need to be copied to accounts.
Added code to AbstractAccountsProvider.setCurrentAccount to copy old preferences to the current account.

PiperOrigin-RevId: 199840009